[DM-23081] Copy the example-tap code from CADC
This is where we're starting from.  First, let's take the
example-tap directory out of https://github.com/opencadc/tap

We need to take this as a copy because this is what builds the
TAP .war file and has all the parts that we might want to
configure, such as the results store, or the job manager.