[DM-23189] Add in a uws database
Okay, first off, we add a docker/uws directory, that builds the uws
backend DB just like the lsst-tap-service.  This is just a straight
copy.  Add building of the container to build.sh and push.sh .

Remove the SampleJobManager, and switch over to the JobManager that
persists things in postgres in the container mentioned above.

Configure context and web xml files to put use the new classes and
configure the connection to the uws db.