[DM-23421] Inject obscore data in
So what I'm doing here is just getting the data imported.  First,
the data is a compressed CSV file that has been checked in.  It is
imported in 1007 where I use COPY FROM to load it from the CSV file.
This required some additional changes to the tap_schema parts to add
more columns.

Also, init-content.sh needed to just execute .sql files so it doesn't
try and execute my compressed archive.  Get rid of the old obscore
placeholder data, as it will just be confusing.