[DM-23421] Make s_region work with pgsphere
So, first off, add the PgsphereRegionConverter to the AdqlQueryImpl.
This means we will work on rewriting the queries to pgsphere compatible
queries.

Then, edit the obscore table so that s_region is not a var string but
instead an spoly (which happens to be a string but is a pgsphere type).
This allows pgsphere to read the s_region column.

Turn on pgsphere for the database in 0000_users.sql

Finally, take the obscore table data and convert polygons to pgsphere
compatible format strings.