[DM-38110] QueryRunner support for tap schema
QueryRunner didn't allow for the tap_schema database to be different
from the main database like the previous version of the tap service,
used in the main tap server.

So let's try to get it.  This takes the QueryRunner from the main
tap service, and tells JobManager to use it.  Also updates context
to allow for more configuration, which will be required in phalanx.