[DM-39825] Try to get TAP UPLOAD to work
In the PluginFactory, it used the "DefaultUploadManager," but even
when that is called, it will say that "UPLOAD is not supported".

That's a bit tricky because technically it is supported in the
capabilities, but this is a message that is emitted by the Upload
Manager.  What we really need is the "BasicUploadManager."

In terms of the name of the account used for the TAP UPLOADS,
Pat has suggested using the same account as for reads.  Otherwise
I can change it back and use some of the .xml files to create
a new named pool.  But it doesn't seem required at this point.